refactor: remove nested components in CharmActionsPanel
