+++ src/panels/ActionsPanel/CharmActionsPanel/CharmActionsPanel.tsx
@@ -1,17 +1,11 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import { Button, ConfirmationModal } from "@canonical/react-components";
+import { Button } from "@canonical/react-components";
 import { useCallback, useMemo, useRef, useState } from "react";
-import reactHotToast from "react-hot-toast";
 import { useSelector } from "react-redux";
-import { useParams } from "react-router-dom";
-import usePortal from "react-useportal";
 
 import LoadingHandler from "components/LoadingHandler";
 import Panel from "components/Panel";
 import RadioInputBox from "components/RadioInputBox";
-import ToastCard from "components/ToastCard";
-import useAnalytics from "hooks/useAnalytics";
-import { useExecuteActionOnUnits } from "juju/api-hooks";
 import ActionOptions from "panels/ActionsPanel/ActionOptions";
 import type {
   ActionOptionValue,
@@ -26,7 +20,7 @@ import {
   getSelectedCharm,
 } from "store/juju/selectors";
 
-import { Label } from "./types";
+import ConfirmationDialog from "./ConfirmationDialog";
 
 const filterExist = <I,>(item: I | null): item is I => !!item;
 
@@ -39,9 +33,6 @@ export default function CharmActionsPanel({
   charmURL,
   onRemovePanelQueryParams,
 }: Props): JSX.Element {
-  const sendAnalytics = useAnalytics();
-  const { userName, modelName } = useParams();
-  const { Portal } = usePortal();
   const [disableSubmit, setDisableSubmit] = useState<boolean>(true);
   const [confirmType, setConfirmType] = useState<ConfirmTypes>(null);
   const [selectedAction, setSelectedAction] = useState<string>();
@@ -49,7 +40,6 @@ export default function CharmActionsPanel({
 
   const selectedApplications = useSelector(getSelectedApplications(charmURL));
   const selectedCharm = useSelector(getSelectedCharm(charmURL));
-  const executeActionOnUnits = useExecuteActionOnUnits(userName, modelName);
   const actionData = useMemo(
     () => selectedCharm?.actions?.specs || {},
     [selectedCharm],
@@ -58,47 +48,6 @@ export default function CharmActionsPanel({
     (total, app) => total + (app["unit-count"] || 0),
     0,
   );
-
-  const executeAction = () => {
-    sendAnalytics({
-      category: "ApplicationSearch",
-      action: "Run action (final step)",
-    });
-
-    if (!selectedAction) return;
-    executeActionOnUnits(
-      // transform applications to unit list for the API
-      selectedApplications
-        .map((a) =>
-          Array(a["unit-count"])
-            .fill("name" in a ? a.name : null)
-            .filter(Boolean)
-            .map((unit, i) => `${unit}-${i}`),
-        )
-        .flat(),
-      selectedAction,
-      actionOptionsValues.current[selectedAction],
-    )
-      .then((payload) => {
-        const error = payload?.actions?.find((e) => e.error);
-        if (error) {
-          throw error;
-        }
-        reactHotToast.custom((t) => (
-          <ToastCard toastInstance={t} type="positive">
-            {Label.ACTION_SUCCESS}
-          </ToastCard>
-        ));
-        return;
-      })
-      .catch(() => {
-        reactHotToast.custom((t) => (
-          <ToastCard toastInstance={t} type="negative">
-            {Label.ACTION_ERROR}
-          </ToastCard>
-        ));
-      });
-  };
 
   const handleSubmit = () => {
     setConfirmType(ConfirmType.SUBMIT);
@@ -135,43 +84,6 @@ export default function CharmActionsPanel({
     },
     [actionData],
   );
-
-  const generateConfirmationModal = () => {
-    if (confirmType && selectedAction) {
-      // Allow for adding more confirmation types, like for cancel
-      // if inputs have been changed.
-      if (confirmType === "submit") {
-        const unitCount = selectedApplications.reduce(
-          (total, app) => total + (app["unit-count"] || 0),
-          0,
-        );
-        // Render the submit confirmation modal.
-        return (
-          <Portal>
-            <ConfirmationModal
-              title={`Run ${selectedAction}?`}
-              cancelButtonLabel={Label.CANCEL_BUTTON}
-              confirmButtonLabel={Label.CONFIRM_BUTTON}
-              confirmButtonAppearance="positive"
-              onConfirm={() => {
-                setConfirmType(null);
-                executeAction();
-                onRemovePanelQueryParams();
-              }}
-              close={() => setConfirmType(null)}
-            >
-              <h4 className="p-muted-heading u-no-margin--bottom">
-                APPLICATION COUNT (UNIT COUNT)
-              </h4>
-              <p data-testid="confirmation-modal-unit-count">
-                {selectedApplications.length} ({unitCount})
-              </p>
-            </ConfirmationModal>
-          </Portal>
-        );
-      }
-    }
-  };
 
   return (
     <Panel
@@ -210,7 +122,14 @@ export default function CharmActionsPanel({
             />
           </RadioInputBox>
         ))}
-        {generateConfirmationModal()}
+        <ConfirmationDialog
+          confirmType={confirmType}
+          selectedAction={selectedAction}
+          selectedApplications={selectedApplications}
+          setConfirmType={setConfirmType}
+          actionOptionsValues={actionOptionsValues}
+          onRemovePanelQueryParams={onRemovePanelQueryParams}
+        />
       </LoadingHandler>
     </Panel>
   );
