+++ src/panels/ActionsPanel/CharmActionsPanel/ConfirmationDialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ConfirmationDialog";
